requirements.txt
openai
requests
beautifulsoup4
textstat
python-dotenv
requests-html
